requirements.txt
django
djangorestframework
Pillow
PyJWT
timedelta